Split algorithm implementation into their own cpp files
